requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langcache>=0.1.0
redisvl>=0.3.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
